moved third party vendor folder from include to lib folder
