+++ test/src/receiver/main.cpp
@@ -14,8 +14,8 @@
 #include <cadmium/logger/common_loggers.hpp>
 
 
-#include "../../../include/vendor/NDTime.hpp"
-#include "../../../include/vendor/iestream.hpp"
+#include "../../../lib/vendor/NDTime.hpp"
+#include "../../../lib/vendor/iestream.hpp"
 
 #include "../../../include/data_structures/message.hpp"
